requirements.txt
streamlit
pandas
numpy
yfinance
requests
textblob
python-dotenv
